requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptocode==0.1
discord.py==2.3.2
dj-database-url==0.5.0
Django==4.2.4
django-admin==2.0.2
django-better-admin-arrayfield==1.4.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
docutils @ file:///private/tmp/docutils-20230831-5903-jytmsf/docutils-0.20.1
et-xmlfile==1.1.0
excel-base==1.0.4
flake8==6.1.0
frozenlist==1.4.0
htmlmin==0.1.12
idna==3.4
isoweek==1.3.3
itypes==1.2.0
korean-lunar-calendar==0.3.1
lxml==4.9.3
MarkupPy==1.14
mccabe==0.7.0
multidict==6.0.4
mutagen==1.47.0
numpy==1.25.1
openpyxl==3.1.2
pandas==2.1.0
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodomex==3.18.0
pyflakes==3.1.0
PyJWT==1.7.1
PyMeeus==0.5.12
PyMuPDF==1.22.5
PyNaCl==1.5.0
pyparsing==3.1.0
PyPDF2==1.27.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==1.0.0
pytz==2021.3
pywatchman==1.4.1
PyYAML==6.0
requests==2.31.0
rsa==4.9
screen==1.0.1
selenium==3.141.0
simplejson==3.17.6
six==1.16.0
Slowloris==0.2.6
socks==0
soupsieve==2.4.1
sqlparse==0.4.4
stopit==1.1.2
tablib==3.5.0
termcolor==1.1.0
TimeConvert==3.0.13
toml==0.10.2
types-PyYAML==6.0.12.11
tzdata==2023.3
tzlocal==5.0.1
uritemplate==3.0.1
urllib3==1.26.16
urlman==2.0.1
websockets==11.0.3
xlrd==1.2.0
XlsxWriter==1.3.7
xlutils==2.0.0
xlwt==1.3.0
yarl==1.9.2
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@66ab0814c4baa2dc79c2dd5287bc0ad61a37c5b9
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz#sha256=bbd382a11746c122b22f636cf0219c73f2f4c5621f24e1cdbab85e6e8aee9bfb
